Fix API receiver regression coverage and restore comment encoding
